requirements.txt
uvloop

pyrogram
TgCrypto
git+https://github.com/Jusidama-Bot/pyromod

aiofiles
python-dotenv